Fix "y" solution coordinate and popup square
The square had a position relative to the whole DIV containing
everything, so a new DIV was created to contain the image and that
appearing square. On the other hand, the Header's height had to be
substracted from the Y coordinates provided by the event, as it
affected the real coordinates.