Create Message component
The component tell the user whether he has selected the correct
character or whether he should keep trying.

This component receives two props from the App component, which are
a state used as a class to change the appearance of the popup message.
On the other hand, the second prop is the message to be displayed.