Fix Timer component
Conditional rendering was used to give the format of time to the current
seconds and minutes states, however, the new approach to giving this
format is using only the seconds state and transform it to minutes by
means of a function.

On the other hand, the isGameOver state from App is used to stop the
timer when its component unmounts, so the timer stops when the player
has found all the characters in the game.